display prompt before deleting

+++ JFSGui/src/main/webapp/app/administration/job/list/administration-job-list.js
@@ -37,24 +37,26 @@
         };
 
         function remove(jobOfferId) {
-            $scope.jobOfferId = jobOfferId;
-            JobService.deleteJobOffer($scope.jobOfferId)
-                .then(function(response) {
-                    $scope.noResults = {};
-                    $scope.noResults.info = false;
-                    $scope.noResults.error = false;
-                    if (response.success) {
-                        $scope.dataLoading = false;
-                        getAll();
-                    }
-                    else {
-                        console.error(response);
-                        $scope.noResults.error = true;
-                        $scope.noResults.title = "An error occurred!";
-                        $scope.noResults.text = "Please try again later.";
-                        $scope.dataLoading = false;
-                    }
-                });
+            if(confirm("Are you sure you want to delete the job offer?")) {
+                $scope.jobOfferId = jobOfferId;
+                JobService.deleteJobOffer($scope.jobOfferId)
+                    .then(function (response) {
+                        $scope.noResults = {};
+                        $scope.noResults.info = false;
+                        $scope.noResults.error = false;
+                        if (response.success) {
+                            $scope.dataLoading = false;
+                            getAll();
+                        }
+                        else {
+                            console.error(response);
+                            $scope.noResults.error = true;
+                            $scope.noResults.title = "An error occurred!";
+                            $scope.noResults.text = "Please try again later.";
+                            $scope.dataLoading = false;
+                        }
+                    });
+            }
         };
 
         function search() {
